+++ src/runtime_src/xdp/profile/database/events/vtf_event.cpp
@@ -27,7 +27,7 @@ namespace xdp {
   // Base class definitions
   // **************************
 
-  VTFEvent::VTFEvent(uint64_t s_id, uint64_t ts, VTFEventType ty) :
+  VTFEvent::VTFEvent(uint64_t s_id, double ts, VTFEventType ty) :
     id(0), start_id(s_id), timestamp(ts), type(ty)
   {
   }
@@ -47,8 +47,8 @@ namespace xdp {
   void VTFEvent::dumpTimestamp(std::ofstream& fout)
   {
     // Host events are accurate up to microseconds.
-    //  Timestamps are in milliseconds, so the precision should be 3 past
-    //  the decimal point
+    // Timestamps are in milliseconds, so the precision should be 3 past
+    // the decimal point
     std::ios_base::fmtflags flags = fout.flags() ;
     fout << std::fixed << std::setprecision(6) << (timestamp/1.0e6) ;
     fout.flags(flags) ;
@@ -164,7 +164,7 @@ namespace xdp {
   // API Call definitions
   // **************************
 
-  APICall::APICall(uint64_t s_id, uint64_t ts, uint64_t name, VTFEventType ty)
+  APICall::APICall(uint64_t s_id, double ts, uint64_t name, VTFEventType ty)
          : VTFEvent(s_id, ts, ty),
            functionName(name)
   {
